enhancement: add shallow musl patches on top
Changes were made to make it compile on Alpine Linux. This is, however, only superficial changes without deep understanding of the code itself, therefore not suitable for production.

+++ src/pcem/vid_voodoo_banshee.cpp
@@ -221,9 +221,9 @@ static int banshee_vga_vsync_enabled(banshee_t *banshee)
 static void banshee_update_irqs(banshee_t *banshee)
 {
     if (banshee->vblank_irq > 0 && banshee_vga_vsync_enabled(banshee)) {
-        pci_set_irq(NULL, PCI_INTA);
+        pci_set_irq((intptr_t)NULL, PCI_INTA);
     } else {
-        pci_clear_irq(NULL, PCI_INTA);
+        pci_clear_irq((intptr_t)NULL, PCI_INTA);
     }
 }
